system: split users into a new file
So we can re use the users in other services without needing to parse
cfg.users.users. That can also work but there's a chance that things
like the root account or any service account that did not properly mark
isNormalUser can create problems.

This is kind of a pseudo-LDAP.